requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
opcua-asyncio>=1.0.0
cryptography>=37.0.0
werkzeug>=2.0.0
mariadb>=1.0.0